requirements.txt
# 界面相关依赖
gradio>=4.0.0
numpy>=1.24.0

# 视觉感知模块
torch==2.1.0
torchvision>=0.15.0
transformers>=4.30.0
opencv-python>=4.8.0
Pillow>=10.0.0
